# food-retail-marketplace/ardo-backend/docs/swagger.yaml
definitions:
...
  management_catalog.Catalog:
    properties:
      createdAt:
        type: string
      id:
        type: string
      promo:
        items:
          $ref: '#/definitions/management_catalog.Section'
        type: array
      structure:
        items:
          $ref: '#/definitions/management_catalog.Section'
        type: array
      updated_at:
        type: string
      version:
        type: integer
    type: object
  management_catalog.CatalogsData:
    properties:
      catalogs:
        items:
          $ref: '#/definitions/management_catalog.Catalog'
        type: array
      count:
        type: integer
    type: object
  management_catalog.Category:
    properties:
      categoryId:
        type: string
      products:
        items:
          $ref: '#/definitions/management_catalog.Product'
        type: array
    type: object
  management_catalog.Product:
    properties:
      productId:
        type: string
    type: object
  management_catalog.Section:
    properties:
      categories:
        items:
          $ref: '#/definitions/management_catalog.Category'
        type: array
      sectionId:
        type: string
    type: object
  management_catalog.TimeOfPublish:
    properties:
      publishedAt:
        type: string
    type: object
  management_catalog.UpdateCatalogIn:
    properties:
      promo:
        items:
          $ref: '#/definitions/management_catalog.Section'
        type: array
      structure:
        items:
          $ref: '#/definitions/management_catalog.Section'
        type: array
    type: object
...
  response.Error:
    properties:
      messages:
        example:
        - error message
        items:
          type: string
        type: array
      success:
        example: false
        type: boolean
      traceId:
        example: string
        type: string
    type: object
  response.Success:
    properties:
      data: {}
      success:
        example: true
        type: boolean
      traceId:
        example: string
        type: string
    type: object
host: api.ardogroup.org
info:
  contact:
    name: API Support
    url: https://nassyrov.t.me
  description: Detailed info about all endpoints
  title: Ardo Backend API
  version: 1.0.0
paths:
...
  /management/catalog:
    get:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Success'
            - properties:
                data:
                  $ref: '#/definitions/management_catalog.CatalogsData'
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Get all catalogs
      tags:
      - Management Catalog
    post:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Create catalog
      tags:
      - Management Catalog
  /management/catalog/{catalog_id}:
    get:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Success'
            - properties:
                data:
                  $ref: '#/definitions/management_catalog.Catalog'
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Get catalog by id
      tags:
      - Management Catalog
    put:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      parameters:
      - description: Update catalog object
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/management_catalog.UpdateCatalogIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Update catalog
      tags:
      - Management Catalog
...
  /management/publish-time/{catalog_id}:
    get:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Success'
            - properties:
                data:
                  $ref: '#/definitions/management_catalog.TimeOfPublish'
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Get time of last publish
      tags:
      - Management Catalog
  /management/publish/{catalog_id}:
    post:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Publish catalog
      tags:
      - Management Catalog
...
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Cookie
    type: apiKey
swagger: "2.0"